requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apturl==0.5.2
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.4
command-not-found==0.3
cryptography==46.0.5
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
et_xmlfile==2.0.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.133.0
filelock==3.20.0
fsspec==2025.12.0
google-api-core==2.30.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.65.0
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.2.0
olefile==0.46
openai==2.21.0
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pexpect==4.8.0
pillow==12.1.1
proto-plus==1.27.1
protobuf==5.29.6
ptyprocess==0.7.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycairo==1.20.1
pycparser==3.0
pycups==2.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyodbc==5.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2026.2.19
reportlab==3.6.8
requests==2.32.5
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
SecretStorage==3.3.1
sentence-transformers==5.2.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.52.1
sympy==1.14.0
systemd-python==234
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0+cpu
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==1.26.5
uvicorn==0.41.0
wadllib==1.3.6
websockets==16.0
xdg==5
xkit==0.0.0
zipp==1.0.0
